Revert "Add grub and grubby"
This reverts commit b9a31f21510ad5e0f689b5e11cd26c2cfb43cb58.

scilinux only has grub0.97 available, but we currently install/use grub2
in the disk. So we can't do the dynamic grub generation because grub2
want's to be called from within the chroot and later on if the client
changes the kernel the packages won't know to call `grub-mkconfig`. We
may be able to use grubby as the kernel cmdline copier that
`grub-mkconfig` is basically used for, but we can cross that bridge when
someone wants to change scilinux kernels (probably no one, ever).