Allow image to be extracted to temp directory
From the original commented line, it seems the intent had been to extract the image to the temp directory instead of the working directory but encountered issues with `sed` misinterpreting the un-escaped slashes in the absolute path returned by `mktemp`. By using an alternate sed delimiter (the colon) there should be no characters to escape in the path returned from `mktemp`.

Strictly there is likely a safer way to fix this without using `sed` at all but this is the least-code-touched solution.